Telegram_Bot/requirements.txt
APScheduler==3.6.3
argon2-cffi==20.1.0
async-generator==1.10
attrs==20.3.0
autograd==1.3
autograd-gamma==0.5.0
backcall==0.2.0
beautifulsoup4==4.9.3
bitso-py==3.0.1
bleach==3.2.1
Brotli==1.0.9
bs4==0.0.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
chart-studio==1.1.0
click==7.1.2
colorama==0.4.4
colorlover==0.3.0
cryptography==3.3.1
cufflinks==0.17.3
cycler==0.10.0
dash==1.19.0
dash-core-components==1.15.0
dash-html-components==1.1.2
dash-renderer==1.9.0
dash-table==4.11.2
decorator==4.4.2
defusedxml==0.6.0
distro==1.5.0
dnspython==1.16.0
entrypoints==0.3
et-xmlfile==1.0.1
feature-engine==1.0.2
Flask==1.1.2
Flask-Compress==1.8.0
future==0.18.2
h5py==2.10.0
idna==2.10
ipykernel==5.3.4
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.17.2
Jinja2==2.11.2
joblib==0.17.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.1
jupyter-core==4.7.0
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.6
jupyter-nbextensions-configurator==0.4.1
jupyterlab-pygments==0.1.2
jupyterthemes==0.20.0
Keras==2.4.3
kiwisolver==1.3.1
lesscpy==0.14.0
lifelines==0.25.7
lxml==4.5.2
MarkupSafe==1.1.1
matplotlib==3.3.3
mistune==0.8.4
mock==4.0.3
msgpack==0.5.6
multitasking==0.0.9
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
nltk==3.5
notebook==6.1.5
numpy==1.19.1
openpyxl==3.0.5
packaging==20.4
pandas==1.1.1
pandas-datareader==0.9.0
pandocfilters==1.4.3
parso==0.7.1
patsy==0.5.1
pickleshare==0.7.5
Pillow==8.0.1
plotly==4.14.3
ply==3.11
praw==7.2.0
prawcore==2.0.0
prometheus-client==0.9.0
prompt-toolkit==3.0.8
pycparser==2.20
Pygments==2.7.2
pymongo==3.11.0
PyMySQL==0.10.1
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python-telegram-bot==13.1
pyti==0.2.2
pytz==2020.1
pywin32==300
pywinpty==0.5.7
PyYAML==5.4.1
pyzmq==20.0.0
qrcode==6.1
qtconsole==5.0.0
QtPy==1.9.0
regex==2020.11.13
requests==2.24.0
retrying==1.3.3
scikit-learn==0.23.2
scipy==1.5.2
seaborn==0.11.0
selenium==3.141.0
Send2Trash==1.5.0
six==1.15.0
soupsieve==2.1
spotipy==2.16.1
SQLAlchemy==1.3.20
statsmodels==0.12.1
tabula-py==2.2.0
terminado==0.9.1
testpath==0.4.4
threadpoolctl==2.1.0
toolz==0.11.1
tornado==6.1
tqdm==4.57.0
traitlets==5.0.5
tzlocal==2.1
update-checker==0.18.0
urllib3==1.25.10
uuid==1.30
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==1.0.1
widgetsnbextension==3.5.1
xlrd==1.2.0
yaspin==1.0.0
yellowbrick==1.2
yfinance==0.1.54
